fix: renamed loadObjects to load_objects
